test: Refactor UserRepository testCode - 테스트에서 H2 데이터 베이스 사용하도록 수정
