+++ build.gradle
@@ -30,6 +30,7 @@ dependencies {
 	// DB
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'com.mysql:mysql-connector-j'
+	implementation 'com.h2database:h2'
 
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 	implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
@@ -52,4 +53,7 @@ dependencies {
 
 tasks.named('test') {
 	useJUnitPlatform()
+
+	// 테스트 실행 후 애플리케이션 계속 실행 유지
+	finalizedBy 'bootRun'
 }
