+++ src/main/java/com/eunsun/travel_mate/domain/User.java
@@ -19,7 +19,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-@Table(name = "user")
+@Table(name = "`user`")
 public class User extends Base{
 
   @Id
